Add lattice wrapper around grid graph constructor. Reorganize code into more files.
